poniendo el turno del jugador y lo siguiente es cambiar la imagen cuando se pincha en la carta

+++ src/main/java/com/mycompany/acertijosmatematicos/Logica.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class Logica {
     
     static String [][] cuadricula = new String[Tablero.anchoTablero][Tablero.altoTablero];
-    static int [][] cuadriculaResp = new int[Tablero.anchoTablero][Tablero.altoTablero];
+    static double [][] cuadriculaResp = new double[Tablero.anchoTablero][Tablero.altoTablero];
     
     static char turnoJugador = '1';
     final char JUGADOR_1 = '1';
